Support Sign in with Slack tokens (#31)
If you use "Sign in with Slack" then the token recieved can't access the
"auth.test" endpoint. Instead the identity needs to be fetched from the
"users.identity" endpoint.

This change makes the slack strategy fetch the identity in addition to
the auth wheth the "identity.basic" scope is present.

Note that the strategy always attempts to fetch the auth because
sometimes the token can access _both_ endpoints (say if the scope has
"users:read" _and_ "identity.basic"). The difference is that if the
"identity.basic" scope is present we don't consider an "invalid_auth"
response from the "auth.test" endpoint to be an error case.

ref: https://api.slack.com/docs/sign-in-with-slack